+++ include/headers/evaluate_position.hpp
@@ -5,5 +5,5 @@
 #include "board.hpp"
 
 namespace chess::engine {
-    double evaluate_position(Board& board, uint8_t color);
+    int evaluate_position(Board& board, uint8_t color);
 }
